requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg[c]==3.1.18
drf-spectacular==0.27.2
Pillow==10.3.0
